# specs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Scopegate API contract",
    "version": "1.0.0",
    "description": "Executable local product contract. Authority and trust boundaries are defined in docs/SECURITY.md."
  },
  "servers": [
    {
      "url": "http://localhost:5187",
      "description": "Local same-origin workspace"
    },
    {
      "url": "http://localhost:8457",
      "description": "Local API transport"
    }
  ],
  "paths": {
...
    "/api/v1/organizations/{organization_id}/projects/{project_id}/resources": {
      "get": {
        "operationId": "listProjectResources",
        "summary": "listProjectResources",
        "tags": [
          "session"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourcePage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "408": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Query"
          },
          {
            "$ref": "#/components/parameters/Locale"
          },
          {
            "$ref": "#/components/parameters/View"
          }
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/projects/{project_id}/entitlements/{resource_id}": {
      "put": {
        "operationId": "setEntitlement",
        "summary": "setEntitlement",
        "tags": [
          "platform"
        ],
        "security": [
          {
            "platformBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResource"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "412": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "408": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/ResourceId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/EntitlementMatch"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntitlementState"
              }
            }
          }
        },
        "description": "Platform authority required. Recompute impact token under ordered resource/share then organization/exclusive locks. Stale revision or impact returns 412 without effects. Disable atomically revokes all scoped active grants, increments each changed membership version once and revokes whole pending invitations containing this pair. Reenable grants nothing. Exceeding the impact cap returns 409 before writes."
      }
    },
...
    "/api/v1/catalog/resources/{external_key}": {
      "put": {
        "operationId": "publishResource",
        "summary": "publishResource",
        "tags": [
          "publisher"
        ],
        "security": [
          {
            "publisherBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "408": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/ExternalKey"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogPublish"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/organizations/{organization_id}/projects/{project_id}/entitlements/{resource_id}/impact": {
      "get": {
        "operationId": "getEntitlementImpact",
        "summary": "Preview bounded entitlement transition impact",
        "tags": [
          "platform"
        ],
        "security": [
          {
            "platformBearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/ResourceId"
          },
          {
            "name": "desired_status",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "disabled"
              ]
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "200": {
            "description": "Scoped impact and concurrency fingerprint; grants no authority.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitlementImpact"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "408": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "scopegate_session"
      },
      "publisherBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Validated publisher machine credential; catalog write only"
      },
      "platformBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Validated platform operations credential; administrative provisioning only, no implicit resource use"
      }
    },
    "parameters": {
      "OrganizationId": {
        "name": "organization_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "ProjectId": {
        "name": "project_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "ResourceId": {
        "name": "resource_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "ExternalKey": {
        "name": "external_key",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[A-Za-z0-9_-]{1,128}$"
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "schema": {
          "type": "string",
          "maxLength": 512
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 25
        }
      },
      "Query": {
        "name": "q",
        "in": "query",
        "schema": {
          "type": "string",
          "maxLength": 200
        }
      },
      "Locale": {
        "name": "locale",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": [
            "en",
            "pt",
            "es",
            "fr"
          ],
          "default": "en"
        }
      },
      "View": {
        "name": "view",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": [
            "granted",
            "entitled"
          ],
          "default": "granted"
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 128
        }
      },
...
      "EntitlementMatch": {
        "name": "If-Match",
        "in": "header",
        "required": true,
        "description": "Entitlement revision; v0 only when no entitlement exists.",
        "schema": {
          "type": "string",
          "pattern": "^\"v[0-9]+\"$"
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "correlation_id": {
                "type": "string",
                "format": "uuid"
              },
              "fields": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "required": [
              "code",
              "message",
              "correlation_id"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Localization": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "locale": {
            "type": "string",
            "enum": [
              "en",
              "pt",
              "es",
              "fr"
            ]
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 300
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 10000
          },
          "tags": {
            "type": "array",
            "maxItems": 30,
            "uniqueItems": true,
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            }
          }
        },
        "required": [
          "locale",
          "title",
          "tags"
        ]
      },
      "Resource": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "external_key": {
            "type": "string",
            "minLength": 1
          },
          "catalog_version": {
            "type": "integer",
            "minimum": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "published",
              "archived"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "locale": {
            "type": [
              "string",
              "null"
            ]
          },
          "fallback_used": {
            "type": "boolean"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "external_key",
          "catalog_version",
          "status",
          "fallback_used"
        ]
      },
      "CatalogPublish": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 0
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "published",
              "archived"
            ]
          },
          "avatar_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "localizations": {
            "type": "array",
            "minItems": 1,
            "maxItems": 4,
            "items": {
              "$ref": "#/components/schemas/Localization"
            }
          }
        },
        "required": [
          "expected_version",
          "version",
          "status",
          "localizations"
        ]
      },
      "ProjectResource": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "resource_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled"
            ]
          },
          "entitlement_version": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "project_id",
          "resource_id",
          "status",
          "entitlement_version"
        ]
      },
      "EntitlementState": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled"
            ]
          },
          "impact_token": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          }
        },
        "required": [
          "status",
          "impact_token",
          "reason"
        ]
      },
...
      "ResourcePage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Resource"
            },
            "maxItems": 100
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "next_cursor"
        ]
      },
...
      "EntitlementImpact": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "entitlement_version": {
            "type": "integer",
            "minimum": 0
          },
          "desired_status": {
            "type": "string",
            "enum": [
              "active",
              "disabled"
            ]
          },
          "affected_memberships": {
            "type": "integer",
            "minimum": 0,
            "maximum": 200
          },
          "active_grants": {
            "type": "integer",
            "minimum": 0,
            "maximum": 500
          },
          "pending_invitations": {
            "type": "integer",
            "minimum": 0,
            "maximum": 200
          },
          "impact_token": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          }
        },
        "required": [
          "entitlement_version",
          "desired_status",
          "affected_memberships",
          "active_grants",
          "pending_invitations",
          "impact_token"
        ]
      },
...
    },
    "responses": {
      "Error": {
        "description": "Stable machine-readable error. Inaccessible objects use 404; known insufficient management scope uses 403.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}